# server/common/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: myapp
  description: myapp description
host: 127.0.0.1:3000
schemes:
  - http
basePath: /api
...
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /v1/auth/update-password:
    put:
      tags:
        - Auth
      description: Update user password
      operationId: updatePassword
      parameters:
        - name: user
          in: body
          description: User data
          required: true
          schema: 
            $ref: "#/definitions/UpdatePasswordRequest"
      responses:
        204:
          description: Empty Response
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Not authorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbbiden"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Forbbiden"
          schema:
            $ref: "#/definitions/Error" 
...
  /v1/auth/check:
    get:
      tags:
        - Auth
      description: Check token
      operationId: checkToken
      parameters:
        - in: header
          name: Authorization
          type: string
          required: true
      responses:
        204:
          description: Valid Token, Body is empty
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Not authorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbbiden"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Forbbiden"
          schema:
            $ref: "#/definitions/Error" 
...
definitions:
  UpdatePasswordRequest:
    type: object
    title: Update Password request
    required:
      - userId
      - oldPassword
      - newPassword
    properties:
      userId:
        type: string
        pattern: /^[0-9a-fA-F]{24}$/
        description: User's ID
      oldPassword: 
        type: string
        minLength: 6
        maxLength: 12
        description: User's old password
      newPassword:
        type: string
        minLength: 6
        maxLength: 12
        description: New password
...
  Error:
    type: object
    title: Error details
    properties:
      code:
        type: string
        description: Business error code description
      errors:
        type: array
        items:
          type: string
          description: Error description readable for humans